+++ cmd/jiraissuecalculator/BUILD.bazel
@@ -8,7 +8,7 @@ go_library(
     visibility = ["//visibility:private"],
     deps = [
         "//cmd/jiraissuecalculator/calculator",
-        "//pkg/configrepository",
+        "//pkg/config/repo",
         "//pkg/jira/repo/calculationsrepository",
         "//pkg/jira/repo/issuerepository",
         "//vendor/github.com/Netflix/go-env",
