+++ pkg/config/models/BUILD.bazel
@@ -1,0 +1,8 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "models",
+    srcs = ["models.go"],
+    importpath = "github.com/stebennett/squad-dashboard/pkg/config/models",
+    visibility = ["//visibility:public"],
+)
