+++ pkg/config/models/models.go
@@ -1,0 +1,9 @@
+package models
+
+type Config struct {
+	JiraProject          string
+	JiraToDoStates       []string
+	JiraInProgressStates []string
+	JiraDoneStates       []string
+	NonWorkingDays       []string
+}
