--- pkg/configloader/BUILD.bazel
@@ -1,9 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "configloader",
-    srcs = ["configloader.go"],
-    importpath = "github.com/stebennett/squad-dashboard/pkg/configloader",
-    visibility = ["//visibility:public"],
-    deps = ["//pkg/configrepository"],
-)
